Remove old form builer helpers
Now we've completed the migration we can get rid of the old form builder helper, and make the `GOVUKDesignSystemFormBuilder::FormBuilder` the default builder in the initializer.

Cleaned up as well the "workaround" for old error messages as this is not needed anymore. All errors are using the new builder.
